add if(isset($_SESSION['user']))

+++ Public/app/posts/delete.php
@@ -6,14 +6,17 @@ require __DIR__ . '/../autoload.php';
 
 // In this file we delete new posts in the database.
 
-if (isset($_POST['post-id'], $_POST['post-name'])) {
+if (isset($_POST['post-id'], $_POST['post-name'], $_SESSION['user'])) {
     $postId = filter_var($_POST['post-id'], FILTER_SANITIZE_STRING);
     $postName = filter_var($_POST['post-name'], FILTER_SANITIZE_STRING);
-    $statement = $pdo->prepare('DELETE FROM posts where id = :id');
+    $userId = (int) $_SESSION['user']['id'];
+
+    $statement = $pdo->prepare('DELETE FROM posts where id = :id AND user_id = :user_id');
     if (!$statement) {
         die(var_dump($pdo->errorInfo()));
     }
     $statement->bindParam(':id', $postId, PDO::PARAM_STR);
+    $statement->bindParam(':user_id', $userId, PDO::PARAM_STR);
     $statement->execute();
 
     $path = '../../uploads/' . $postName;
